requirements.txt
chromadb==0.4.22
google-generativeai==0.8.5
python-dotenv==1.0.0
